+++ src/screens/week.js
@@ -7,9 +7,7 @@ const WeekStack = createStackNavigator({
 },
 {
   initialRouteName: 'weekView',
-  defaultNavigationOptions: {
-    ...headerOptions,
-  },
+  ...headerOptions,
 });
 
 export default WeekStack;